refactor: calendly user dto and the mapper for it
Changes:
- Move all calendly-related types to the types lib
- Some dev notes about the "md.d.ts" file
