+++ apps/dariaklover/src/app/types/md.d.ts
@@ -1,1 +1,2 @@
+// TODO: Remove it from "types"
 declare module '*.md';
